backend/requirements.txt
# ============================================
# Core Framework
# ============================================
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1

# ============================================
# LangChain + RAG
# ============================================
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-chroma==0.2.0
langchain-huggingface==0.1.2

# ============================================
# Vector Store
# ============================================
chromadb==0.5.23

# ============================================
# Embeddings (Free - Local)
# ============================================
sentence-transformers==3.3.1

# ============================================
# Data & APIs
# ============================================
requests==2.32.3
finnhub-python==2.4.20
pandas==2.2.3
numpy==1.26.4

# ============================================
# Sentiment Analysis (Free)
# ============================================
vaderSentiment==3.3.2

# ============================================
# Utilities
# ============================================
python-dateutil==2.9.0
cachetools==5.5.1